requirements.txt
asgiref==3.3.1
Django==3.1.7
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.2
pytz==2021.1
sqlparse==0.4.1
